requirements.txt
adal==0.4.5
appdirs==1.4.3
asn1crypto==0.22.0
azure==2.0.0
azure-batch==3.0.0
azure-common==1.1.6
azure-datalake-store==0.0.9
azure-graphrbac==0.30.0
azure-keyvault==0.3.3
azure-mgmt==1.0.0
azure-mgmt-authorization==0.30.0
azure-mgmt-batch==4.0.0
azure-mgmt-cdn==0.30.3
azure-mgmt-cognitiveservices==1.0.0
azure-mgmt-compute==1.0.0
azure-mgmt-containerregistry==0.2.1
azure-mgmt-datalake-analytics==0.1.4
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.1.4
azure-mgmt-devtestlabs==2.0.0
azure-mgmt-dns==1.0.1
azure-mgmt-documentdb==0.1.3
azure-mgmt-iothub==0.2.2
azure-mgmt-keyvault==0.31.0
azure-mgmt-logic==2.1.0
azure-mgmt-monitor==0.2.1
azure-mgmt-network==1.0.0
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==0.1.0
azure-mgmt-redis==4.1.0
azure-mgmt-resource==1.1.0
azure-mgmt-scheduler==1.1.2
azure-mgmt-sql==0.5.1
azure-mgmt-storage==1.0.0
azure-mgmt-trafficmanager==0.30.0
azure-mgmt-web==0.32.0
azure-nspkg==2.0.0
azure-servicebus==0.21.1
azure-servicefabric==5.6.130
azure-servicemanagement-legacy==0.20.6
azure-storage==0.34.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.3
cryptography==1.8.2
dj-database-url==0.4.2
Django==1.11.1
django-storages==1.5.2
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
gunicorn==19.7.1
idna==2.5
isodate==0.5.4
keyring==10.3.2
msrest==0.4.8
msrestazure==0.4.7
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
requests==2.16.5
requests-oauthlib==0.8.0
SecretStorage==2.3.1
six==1.10.0
urllib3==1.21.1
whitenoise==3.3.0
